Shift /users APIs to /auth
