+++ hanki-backend/src/main/java/com/hanki/backend/security/SecurityConfig.java
@@ -34,7 +34,7 @@ public class SecurityConfig {
                 // disable CSRF
                 .csrf(customizer -> customizer.disable())
                 // TODO - Update the exact endpoint(s) that don't require auth
-                .authorizeHttpRequests(request -> request.requestMatchers("/users/register", "/users/login").permitAll())
+                .authorizeHttpRequests(request -> request.requestMatchers("/auth/register", "/auth/login").permitAll())
                 // all other endpoints require auth
                 .authorizeHttpRequests(request -> request.anyRequest().authenticated())
                 // to allow requests via browser
